requirements.txt
Flask==1.1.2
Pillow==7.2.0
Numpy==1.19.2
torch==1.4.0
torchvision==0.5.0
dominate==2.4.0
visdom==0.1.8.8